usb: split HID into per-class files and unify all USB state under grid_usb_state
Split grid_usb_hid.c into grid_usb_keyboard.c, grid_usb_mouse.c,
grid_usb_macro.c, grid_usb_gamepad.c. The HID layer becomes a
coordinator: grid_usb_hid_on_connect/disconnect delegates to each
sub-class lifecycle hook.

Introduce struct grid_usb_hid_model with keyboard, mouse, macro, and
gamepad embedded as value members. Lift this further into struct
grid_usb_model (acm + midi + hid) with a single global grid_usb_state
defined in grid_usb.c. All individual per-class globals are removed;
callers access grid_usb_state.acm, grid_usb_state.midi,
grid_usb_state.hid.{keyboard,mouse,macro,gamepad}.

grid_usb.h becomes the USB umbrella header; including it gives access
to all sub-class types and the single global state.
