+++ common/src/c/grid_port.c
@@ -8,9 +8,7 @@
 #include "grid_decode.h"
 #include "grid_platform.h"
 #include "grid_protocol.h"
-#include "grid_usb_acm.h"
-
-extern bool grid_usb_connected(void);
+#include "grid_usb.h"
 
 extern struct grid_decoder_collection* grid_decoder_to_ui_reference;
 extern struct grid_decoder_collection* grid_decoder_to_usb_reference;
@@ -245,7 +243,7 @@ void grid_port_send_usb(struct grid_port* port) {
 
   assert(port->type == GRID_PORT_USB);
 
-  if (grid_usb_acm_dtr(&grid_usb_acm_state) && !grid_usb_acm_tx_ready(&grid_usb_acm_state)) {
+  if (grid_usb_acm_dtr(&grid_usb_state.acm) && !grid_usb_acm_tx_ready(&grid_usb_state.acm)) {
     return;
   }
 
@@ -260,7 +258,7 @@ void grid_port_send_usb(struct grid_port* port) {
   }
 
   grid_port_decode_msg(grid_decoder_to_usb_reference, &msg);
-  grid_usb_acm_write(&grid_usb_acm_state, msg.data, msg.length);
+  grid_usb_acm_write(&grid_usb_state.acm, msg.data, msg.length);
 }
 
 void grid_port_send_ui(struct grid_port* port) {
