+++ common/src/c/grid_transport.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "grid_usb_acm.h"
+#include "grid_usb.h"
 
 struct grid_transport grid_transport_state;
 
@@ -253,7 +253,7 @@ void grid_transport_rx_broadcast_tx(struct grid_transport* transport, struct gri
       }
 
       if (type == GRID_PORT_USB) {
-        grid_usb_acm_state.tx_dropped++;
+        grid_usb_state.acm.tx_dropped++;
       }
 
       continue;
