+++ common/src/c/grid_usb.c
@@ -9,38 +9,32 @@
  *   grid_usb_hid.c   - HID keyboard model + gamepad + enable/disable
  */
 
+#include "tusb.h"
+
 #include "grid_usb.h"
 
 #include <string.h>
 
 #include "grid_protocol.h"
 #include "grid_swsr.h"
-#include "grid_usb_acm.h"
 #include "grid_usb_desc.h"
-#include "grid_usb_gamepad.h"
-#include "grid_usb_hid.h"
-#include "grid_usb_midi.h"
-#include "grid_usb_mouse.h"
-#include "tusb.h"
+
+struct grid_usb_model grid_usb_state = {0};
 
 bool grid_usb_connected(void) { return tud_mounted(); }
 
 void grid_usb_task(void) { tud_task_ext(0, false); }
 
 void tud_mount_cb(void) {
-  grid_usb_midi_on_connect(&grid_usb_midi_state);
-  grid_usb_hid_on_connect(&grid_macro_state, &grid_keyboard_state);
-  grid_usb_mouse_on_connect(&grid_mouse_state);
-  grid_usb_gamepad_on_connect(&grid_gamepad_state);
-  grid_usb_acm_on_connect(&grid_usb_acm_state);
+  grid_usb_midi_on_connect(&grid_usb_state.midi);
+  grid_usb_hid_on_connect(&grid_usb_state.hid);
+  grid_usb_acm_on_connect(&grid_usb_state.acm);
 }
 
 void tud_umount_cb(void) {
-  grid_usb_midi_on_disconnect(&grid_usb_midi_state);
-  grid_usb_hid_on_disconnect(&grid_macro_state, &grid_keyboard_state);
-  grid_usb_mouse_on_disconnect(&grid_mouse_state);
-  grid_usb_gamepad_on_disconnect(&grid_gamepad_state);
-  grid_usb_acm_on_disconnect(&grid_usb_acm_state);
+  grid_usb_midi_on_disconnect(&grid_usb_state.midi);
+  grid_usb_hid_on_disconnect(&grid_usb_state.hid);
+  grid_usb_acm_on_disconnect(&grid_usb_state.acm);
 }
 
 static tusb_desc_device_t s_device_desc = {
@@ -116,12 +110,9 @@ void grid_usb_init(uint16_t vid, uint16_t pid, const char* serial) {
   s_device_desc.idProduct = pid;
   s_str_table[3] = serial;
   s_device_desc.iSerialNumber = (serial != NULL) ? 0x03 : 0x00;
-  grid_usb_acm_init(&grid_usb_acm_state, GRID_PARAMETER_SPI_TRANSACTION_length * 2);
-  grid_usb_midi_init(&grid_usb_midi_state, GRID_MIDI_TX_BUFFER_SIZE, GRID_MIDI_VOICE_RX_BUFFER_SIZE, GRID_MIDI_SYSEX_BUFFER_SIZE, GRID_MIDI_RTM_BUFFER_SIZE);
-  grid_usb_keyboard_init(&grid_keyboard_state);
-  grid_usb_mouse_init(&grid_mouse_state);
-  grid_usb_macro_init(&grid_macro_state, GRID_MACRO_TX_BUFFER_SIZE, &grid_keyboard_state, &grid_mouse_state);
-  grid_usb_gamepad_init(&grid_gamepad_state, GRID_GAMEPAD_TX_BUFFER_SIZE);
+  grid_usb_acm_init(&grid_usb_state.acm, GRID_PARAMETER_SPI_TRANSACTION_length * 2);
+  grid_usb_midi_init(&grid_usb_state.midi, GRID_MIDI_TX_BUFFER_SIZE, GRID_MIDI_VOICE_RX_BUFFER_SIZE, GRID_MIDI_SYSEX_BUFFER_SIZE, GRID_MIDI_RTM_BUFFER_SIZE);
+  grid_usb_hid_init(&grid_usb_state.hid);
   tusb_init();
 }
