+++ common/src/c/grid_usb_acm.h
@@ -11,8 +11,6 @@ struct grid_usb_acm_model {
   uint32_t tx_dropped;
 };
 
-extern struct grid_usb_acm_model grid_usb_acm_state;
-
 void grid_usb_acm_init(struct grid_usb_acm_model* usb_acm, uint16_t rx_buffer_size);
 void grid_usb_acm_on_connect(struct grid_usb_acm_model* usb_acm);
 void grid_usb_acm_on_disconnect(struct grid_usb_acm_model* usb_acm);
