+++ esp32s3/components/grid_esp32_port/grid_esp32_port.c
@@ -30,10 +30,6 @@
 #include "grid_transport.h"
 #include "grid_ui.h"
 #include "grid_usb.h"
-#include "grid_usb_acm.h"
-#include "grid_usb_gamepad.h"
-#include "grid_usb_hid.h"
-#include "grid_usb_midi.h"
 #include "grid_utask.h"
 
 #include "driver/spi_slave.h"
@@ -234,44 +230,44 @@ struct grid_utask_timer timer_tx_dropped;
 
 void grid_utask_usb_tx(void) {
 
-  if (grid_usb_midi_tx_available(&grid_usb_midi_state)) {
+  if (grid_usb_midi_tx_available(&grid_usb_state.midi)) {
     if (grid_utask_timer_elapsed(&timer_midi_tx)) {
-      grid_usb_midi_tx_flush(&grid_usb_midi_state);
+      grid_usb_midi_tx_flush(&grid_usb_state.midi);
     }
   }
 
-  if (grid_usb_macro_tx_available(&grid_macro_state)) {
+  if (grid_usb_macro_tx_available(&grid_usb_state.hid.macro)) {
     if (grid_utask_timer_elapsed(&timer_keyboard_tx)) {
-      grid_usb_macro_tx_flush(&grid_macro_state);
+      grid_usb_macro_tx_flush(&grid_usb_state.hid.macro);
     }
   }
 
-  if (grid_usb_gamepad_tx_available(&grid_gamepad_state)) {
+  if (grid_usb_gamepad_tx_available(&grid_usb_state.hid.gamepad)) {
     if (grid_utask_timer_elapsed(&timer_gamepad_tx)) {
-      grid_usb_gamepad_tx_flush(&grid_gamepad_state);
+      grid_usb_gamepad_tx_flush(&grid_usb_state.hid.gamepad);
     }
   }
 
-  if (grid_usb_connected() && grid_usb_acm_dtr(&grid_usb_acm_state) && grid_utask_timer_elapsed(&timer_tx_dropped)) {
+  if (grid_usb_connected() && grid_usb_acm_dtr(&grid_usb_state.acm) && grid_utask_timer_elapsed(&timer_tx_dropped)) {
 
-    if (grid_usb_midi_state.tx_dropped > 0) {
-      grid_port_debug_printf("MIDI TX dropped: %lu\n", (unsigned long)grid_usb_midi_state.tx_dropped);
-      grid_usb_midi_state.tx_dropped = 0;
+    if (grid_usb_state.midi.tx_dropped > 0) {
+      grid_port_debug_printf("MIDI TX dropped: %lu\n", (unsigned long)grid_usb_state.midi.tx_dropped);
+      grid_usb_state.midi.tx_dropped = 0;
     }
 
-    if (grid_macro_state.tx_dropped > 0) {
-      grid_port_debug_printf("HID macro TX dropped: %lu\n", (unsigned long)grid_macro_state.tx_dropped);
-      grid_macro_state.tx_dropped = 0;
+    if (grid_usb_state.hid.macro.tx_dropped > 0) {
+      grid_port_debug_printf("HID macro TX dropped: %lu\n", (unsigned long)grid_usb_state.hid.macro.tx_dropped);
+      grid_usb_state.hid.macro.tx_dropped = 0;
     }
 
-    if (grid_gamepad_state.tx_dropped > 0) {
-      grid_port_debug_printf("HID gamepad TX dropped: %lu\n", (unsigned long)grid_gamepad_state.tx_dropped);
-      grid_gamepad_state.tx_dropped = 0;
+    if (grid_usb_state.hid.gamepad.tx_dropped > 0) {
+      grid_port_debug_printf("HID gamepad TX dropped: %lu\n", (unsigned long)grid_usb_state.hid.gamepad.tx_dropped);
+      grid_usb_state.hid.gamepad.tx_dropped = 0;
     }
 
-    if (grid_usb_acm_state.tx_dropped > 0) {
-      grid_port_debug_printf("CDC TX dropped: %lu\n", (unsigned long)grid_usb_acm_state.tx_dropped);
-      grid_usb_acm_state.tx_dropped = 0;
+    if (grid_usb_state.acm.tx_dropped > 0) {
+      grid_port_debug_printf("CDC TX dropped: %lu\n", (unsigned long)grid_usb_state.acm.tx_dropped);
+      grid_usb_state.acm.tx_dropped = 0;
     }
   }
 }
@@ -301,9 +297,9 @@ void grid_utask_midi_rx(struct grid_utask_timer* timer) {
   if (!grid_utask_timer_elapsed(timer)) {
     return;
   }
-  grid_usb_midi_rx_voice_process(&grid_usb_midi_state);
-  grid_usb_midi_rx_sysex_process(&grid_usb_midi_state);
-  grid_usb_midi_rx_rtm_process(&grid_usb_midi_state);
+  grid_usb_midi_rx_voice_process(&grid_usb_state.midi);
+  grid_usb_midi_rx_sysex_process(&grid_usb_state.midi);
+  grid_usb_midi_rx_rtm_process(&grid_usb_state.midi);
 }
 
 extern struct grid_utask_timer timer_draw_event[2];
@@ -565,9 +561,9 @@ void grid_esp32_port_task(void* arg) {
 
     // Duplicate midi rx callback used as a polling mechanism, as the
     // actual callback may not necessarily process all available data
-    grid_usb_midi_rx_poll(&grid_usb_midi_state);
-    grid_usb_acm_rx_poll(&grid_usb_acm_state);
-    grid_usb_acm_rx_process(&grid_usb_acm_state);
+    grid_usb_midi_rx_poll(&grid_usb_state.midi);
+    grid_usb_acm_rx_poll(&grid_usb_state.acm);
+    grid_usb_acm_rx_process(&grid_usb_state.acm);
 
     // Decode for UI
     grid_port_send_ui(port_ui);
